Frontend/requirements.txt
streamlit
Pillow
numpy
opencv-python
psutil
deepface
tf-keras